@@ -1,7 +1,6 @@
-'use strict';
-
-const path = require('path');
-const bin = require('.');
+import path from 'node:path';
+import process from 'node:process';
+import bin from './index.js';
 
 function getProjectRoot() {
   // `projectRoot` on postinstall could be INIT_CWD introduced in npm >= 5.4
