@@ -1,9 +1,7 @@
-'use strict';
-
-const assert = require('assert').strict;
-const binCheck = require('bin-check');
-const hugoBin = require('..');
-const { test } = require('uvu');
+import { strict as assert } from 'node:assert';
+import binCheck from 'bin-check';
+import { test } from 'uvu';
+import hugoBin from '../index.js';
 
 test('should return path to binary and work', () => {
   return binCheck(hugoBin, ['version']).then(works => {
